JAVA file of spin-flip-loss COMSOL model
An integration is performed of molecule flux through regions where loss can happen, weighted by the expected Maxwell-Boltzmann probability for molecules to be at that location. In order for COMSOL to access information about the distribution of the molecules, their ground state Hamiltonian energy levels are encoded analytically as quartic roots.